ids_backend/requirements.txt
django
djangorestframework
torch
numpy
pandas
scikit-learn
gunicorn